requirements.txt
Flask==3.0.0
pandas==2.0.3
mysql-connector-python==8.0.33
psycopg2-binary==2.9.7
openpyxl==3.1.2
gunicorn==21.2.0
numpy==1.24.3
Werkzeug==3.0.1
